// UnrealEngine52/Engine/Source/Programs/Horde/HordeDashboard/src/backend/Api.ts
export type GetPerforceServerStatusResponse = {
	serverAndPort: string;
	baseServerAndPort: string;
	cluster: string;
	numLeases: number;
	status: string;
	detail: string;
}